updated db table to use decimal numbers instead of int

+++ db/schema.sql
@@ -7,6 +7,6 @@ CREATE TABLE fitness
 	id int NOT NULL AUTO_INCREMENT,
 	weight INT NOT NULL,
     exercise varchar(255) NOT NULL,
-	calories INT NOT NULL,
+	calories DECIMAL(4,3) NOT NULL,
 	PRIMARY KEY (id)
 );
